refactor: 패키지 구조 일부 변경
