+++ src/main/java/com/stempo/api/domain/presentation/AuthController.java
@@ -3,7 +3,7 @@ package com.stempo.api.domain.presentation;
 import com.stempo.api.domain.application.service.AuthService;
 import com.stempo.api.domain.presentation.dto.request.AuthRequestDto;
 import com.stempo.api.domain.presentation.dto.response.TokenInfo;
-import com.stempo.api.global.common.dto.ApiResponse;
+import com.stempo.api.global.dto.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletRequest;
