+++ src/main/java/com/stempo/api/domain/presentation/BoardController.java
@@ -5,8 +5,8 @@ import com.stempo.api.domain.domain.model.BoardCategory;
 import com.stempo.api.domain.presentation.dto.request.BoardRequestDto;
 import com.stempo.api.domain.presentation.dto.request.BoardUpdateRequestDto;
 import com.stempo.api.domain.presentation.dto.response.BoardResponseDto;
-import com.stempo.api.global.common.dto.ApiResponse;
-import com.stempo.api.global.common.dto.PagedResponseDto;
+import com.stempo.api.global.dto.ApiResponse;
+import com.stempo.api.global.dto.PagedResponseDto;
 import com.stempo.api.global.exception.InvalidColumnException;
 import com.stempo.api.global.exception.SortingArgumentException;
 import com.stempo.api.global.util.PageableUtil;
