+++ src/main/java/com/stempo/api/domain/presentation/HomeworkController.java
@@ -4,8 +4,8 @@ import com.stempo.api.domain.application.service.HomeworkService;
 import com.stempo.api.domain.presentation.dto.request.HomeworkRequestDto;
 import com.stempo.api.domain.presentation.dto.request.HomeworkUpdateRequestDto;
 import com.stempo.api.domain.presentation.dto.response.HomeworkResponseDto;
-import com.stempo.api.global.common.dto.ApiResponse;
-import com.stempo.api.global.common.dto.PagedResponseDto;
+import com.stempo.api.global.dto.ApiResponse;
+import com.stempo.api.global.dto.PagedResponseDto;
 import com.stempo.api.global.exception.InvalidColumnException;
 import com.stempo.api.global.exception.SortingArgumentException;
 import com.stempo.api.global.util.PageableUtil;
