+++ src/main/java/com/stempo/api/domain/presentation/RecordController.java
@@ -3,7 +3,7 @@ package com.stempo.api.domain.presentation;
 import com.stempo.api.domain.application.service.RecordService;
 import com.stempo.api.domain.presentation.dto.request.RecordRequestDto;
 import com.stempo.api.domain.presentation.dto.response.RecordResponseDto;
-import com.stempo.api.global.common.dto.ApiResponse;
+import com.stempo.api.global.dto.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
