+++ src/main/java/com/stempo/api/domain/presentation/RhythmController.java
@@ -2,7 +2,7 @@ package com.stempo.api.domain.presentation;
 
 import com.stempo.api.domain.application.service.RhythmService;
 import com.stempo.api.domain.presentation.dto.request.RhythmRequestDto;
-import com.stempo.api.global.common.dto.ApiResponse;
+import com.stempo.api.global.dto.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
