+++ src/main/java/com/stempo/api/global/handler/GlobalExceptionHandler.java
@@ -9,8 +9,8 @@ import com.stempo.api.global.auth.exception.AuthenticationInfoNotFoundException;
 import com.stempo.api.global.auth.exception.TokenForgeryException;
 import com.stempo.api.global.auth.exception.TokenNotFoundException;
 import com.stempo.api.global.auth.exception.TokenValidateException;
-import com.stempo.api.global.common.dto.ApiResponse;
-import com.stempo.api.global.common.dto.ErrorResponse;
+import com.stempo.api.global.dto.ApiResponse;
+import com.stempo.api.global.dto.ErrorResponse;
 import com.stempo.api.global.exception.InvalidColumnException;
 import com.stempo.api.global.exception.NotFoundException;
 import com.stempo.api.global.exception.PermissionDeniedException;
